build hooks fixes
