--- packages/build-hooks-monorepo/README.md
@@ -1,1 +1,0 @@
-# @flexn/build-hooks-monorepo
