--- packages/build-hooks-monorepo/package.json
@@ -1,39 +1,0 @@
-{
-    "name": "@rnv/build-hooks-monorepo",
-    "version": "0.38.0-canary.1",
-    "description": "Build Hooks Version manager",
-    "main": "lib/index",
-    "types": "lib/index",
-    "scripts": {
-        "watch": "tsc --watch --preserveWatchOutput",
-        "build": "yarn clean && yarn compile",
-        "clean": "rm -rf ./lib && rm -rf tsconfig.tsbuildinfo",
-        "compile": "tsc -b tsconfig.json"
-    },
-    "files": [
-        "lib",
-        "LICENSE",
-        "src"
-    ],
-    "repository": {
-        "directory": "packages/build-hooks-monorepo",
-        "type": "git",
-        "url": "git+https://github.com/flexn-io/renative.git"
-    },
-    "publishConfig": {
-        "access": "public"
-    },
-    "keywords": [],
-    "author": "Flexn",
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/flexn-io/renative/issues"
-    },
-    "homepage": "https://github.com/flexn-io/renative#readme",
-    "peerDependencies": {
-        "@rnv/core": "*"
-    },
-    "dependencies": {
-        "deepmerge": "3.2.0"
-    }
-}
