--- packages/build-hooks-monorepo/src/__tests__/sanity.test.ts
@@ -1,9 +1,0 @@
-describe('Test', () => {
-    beforeEach(() => {
-        console.log('Testing');
-    });
-    it('init test', async () => {
-        const x = true;
-        expect(x).toEqual(true);
-    });
-});
