--- packages/build-hooks-monorepo/src/index.ts
@@ -1,4 +1,0 @@
-export * from './parser';
-export * from './patcher';
-export * from './synchronizer';
-export * from './versions';
