--- packages/build-hooks-monorepo/src/types.ts
@@ -1,20 +1,0 @@
-export type MonoPackageConfig = {
-    pkgName?: string;
-
-    pkgFile?: any;
-    pkgPath?: string;
-
-    rnvFile?: any;
-    rnvPath?: string;
-
-    metaFile?: any;
-    metaPath?: string;
-
-    plugTempFile?: any;
-    plugTempPath?: string;
-
-    templateConfigFile?: any;
-    templateConfigPath?: string;
-};
-
-export type MonorepoConfig = Record<string, MonoPackageConfig>;
