--- packages/build-hooks-monorepo/tsconfig.json
@@ -1,7 +1,0 @@
-{
-    "extends": "@flexn/typescript-config/tsconfig.lib.node.json",
-    "compilerOptions": {
-        "outDir": "./lib",
-        "rootDir": "./src"
-    }
-}
